removed warning on win

+++ examples_oldgl_glfw/615_PbdClothCad/main.cpp
@@ -63,13 +63,16 @@ void StepTime() {
       dt, gravity, aXYZ, aUVW, aBCFlag);
   dfm2::PBD_TriStrain(
       aXYZt.data(),
-      aXYZt.size() / 3, aETri, aVec2);
+      static_cast<unsigned int>(aXYZt.size() / 3), 
+	  aETri, aVec2);
   dfm2::PBD_Bend(
       aXYZt.data(),
-      aXYZt.size() / 3, aETri, aVec2, 1.0);
+      static_cast<unsigned int>(aXYZt.size() / 3), aETri, aVec2, 1.0);
   dfm2::PBD_Seam(
       aXYZt.data(),
-      aXYZt.size() / 3, aLine.data(), aLine.size() / 2);
+      static_cast<unsigned int>(aXYZt.size() / 3), 
+	  aLine.data(), 
+	  static_cast<unsigned int>(aLine.size() / 2) );
   dfm2::Project_PointsIncludedInBVH_Outside_Cache(
       aXYZt.data(), aInfoNearest,
       aXYZt.size() / 3,
